Refactored evaluation context
