+++ src/humanloop/__init__.py
@@ -27,11 +27,9 @@ from .types import (
     DirectoryWithParentsAndChildrenResponseFilesItem,
     EnvironmentResponse,
     EnvironmentTag,
-    EvaluatedVersionResponse,
     EvaluateeRequest,
     EvaluateeResponse,
     EvaluationEvaluatorResponse,
-    EvaluationReportLogResponse,
     EvaluationResponse,
     EvaluationStats,
     EvaluationStatus,
@@ -86,7 +84,6 @@ from .types import (
     NumericEvaluatorStatsResponse,
     ObservabilityStatus,
     OverallStats,
-    PaginatedDataEvaluationReportLogResponse,
     PaginatedDataEvaluatorResponse,
     PaginatedDataFlowResponse,
     PaginatedDataLogResponse,
@@ -197,8 +194,6 @@ from .requests import (
     EvaluationLogResponseParams,
     EvaluationResponseParams,
     EvaluationStatsParams,
-    EvaluationsDatasetRequestParams,
-    EvaluationsRequestParams,
     EvaluatorActivationDeactivationRequestActivateItemParams,
     EvaluatorActivationDeactivationRequestDeactivateItemParams,
     EvaluatorActivationDeactivationRequestParams,
