+++ src/humanloop/decorators/tool.py
@@ -71,7 +71,7 @@ def tool(
                     output = func(*args, **kwargs)
                     error = None
                 except Exception as e:
-                    logger.error(str(e))
+                    logger.error(f"{func.__name__}: {e}")
                     output = None
                     error = str(e)
